+++ CHANGELOG.md
@@ -8,10 +8,11 @@ All notable changes to this project will be documented in this file.
 
 - Deploy default and support custom affinities ([#391]).
 - Cluster status conditions ([#403])
+- Extend cluster resources for status and cluster operation (paused, stopped) ([#405])
 
 ### Changed
 
-- `operator-rs` `0.31.0` -> `0.37.0` ([#378], [#380], [#400]).
+- `operator-rs` `0.31.0` -> `0.39.0` ([#378], [#380], [#400], [#405]).
 - Bumped stackable image versions to `23.4.0-rc2` ([#378], [#380]).
 - Fragmented `TrinoConfig` ([#379]).
 - Enabled logging and log aggregation ([#380]).
@@ -27,6 +28,7 @@ All notable changes to this project will be documented in this file.
 [#391]: https://github.com/stackabletech/trino-operator/pull/391
 [#400]: https://github.com/stackabletech/trino-operator/pull/400
 [#403]: https://github.com/stackabletech/trino-operator/pull/403
+[#405]: https://github.com/stackabletech/trino-operator/pull/405
 
 ## [23.1.0] - 2023-01-23